refactor(AwardManager): 수익 계산 기능을 YieldManager로 분리
